requirements.txt
Click==7.0
Flask==1.1.1
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
Werkzeug==0.16.0
requests==2.22.0
google-api-python-client==1.12.8
google-auth-oauthlib==0.4.2
pandas==0.25.1